# api-definitions/ai-openapi.json
{
  "paths": {
    "/addresses/ids": {
      "get": {
        "operationId": "addressQueryIDS",
        "tags": [
          "routes"
        ],
        "summary": "Search for an address.",
        "description": "Returns a ranked list of addresses matching the search query in the specified format.",
        "parameters": [
          {
            "in": "query",
            "name": "input",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Specifies the address search string (e.g. 14 Acacia Avenue, Ruislip, HA4 8RG). Can only be blank if rangekm, lat, lon and classificationfilter are supplied in which case the underlyingScore returned is the distance in km."
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "default": 0,
              "maximum": 250
            },
            "description": "Specifies the offset from zero, used for pagination."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "default": 10,
              "maximum": 100
            },
            "description": "Specifies the number of addresses to return."
          },
          {
            "in": "query",
            "name": "classificationfilter",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "description": "Classification code filter. Can be pattern match (ZW*), exact match (RD06), multiple exact match (RD02,RD04) or a preset keyword such as residential, commercial or workplace"
          },
          {
            "in": "query",
            "name": "rangekm",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal"
            },
            "description": "Limit results to those within this number of kilometers of point (decimal e.g. 0.1)"
          },
          {
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal"
            },
            "description": "Latitude of point in decimal format (e.g. 50.705948)."
          },
          {
            "in": "query",
            "name": "lon",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal"
            },
            "description": "Longitude of point in decimal format (e.g. -3.5091076)."
          },
          {
            "in": "query",
            "name": "historical",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": false
            },
            "description": "Include historical addresses"
          },
          {
            "in": "query",
            "name": "matchthreshold",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 5
            },
            "description": "Minimum confidence score (percentage) for match to be included in results."
          },
          {
            "in": "query",
            "name": "verbose",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": false
            },
            "description": "Include the full address details in the response (including relatives, crossRefs, paf and nag)."
          },
          {
            "in": "query",
            "name": "epoch",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "description": "Select a specific AddressBase Epoch to search."
          },
          {
            "in": "query",
            "name": "eboost",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 1
            },
            "description": "Set to 0 to exclude addresses in England"
          },
          {
            "in": "query",
            "name": "nboost",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 1
            },
            "description": "Set to 0 to exclude addresses in Northern Ireland"
          },
          {
            "in": "query",
            "name": "sboost",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 1
            },
            "description": "Set to 0 to exclude addresses in Scotland"
          },
          {
            "in": "query",
            "name": "wboost",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 1
            },
            "description": "Set to 0 to exclude addresses in Wales"
          },
          {
            "in": "query",
            "name": "lboost",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 1
            },
            "description": "Set to 0 to exclude addresses in Channel Islands"
          },
          {
            "in": "query",
            "name": "mboost",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 1
            },
            "description": "Set to 0 to exclude addresses in Isle of Man"
          },
          {
            "in": "query",
            "name": "jboost",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true,
              "format": "decimal",
              "default": 1
            },
            "description": "Set to 0 to exclude addresses outside country boundaries"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. A json return of matched addresses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponseContainerIDS"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Indicates an issue with the request. Further details are provided in the response."
          },
          "401": {
            "description": "Unauthorised. The API key provided with the request is invalid."
          },
          "429": {
            "description": "Server too busy. The Address Index API is experiencing exceptional load."
          },
          "500": {
            "description": "Internal server error. Failed to process the request due to an internal error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressIDS": {
        "properties": {
          "addressEntryId": {
            "type": "string"
          },
          "confidenceScore": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "addressEntryId",
          "confidenceScore"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponseIDS": {
        "properties": {
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressIDS"
            }
          },
          "matchtype": {
            "type": "string"
          },
          "recommendationCode": {
            "type": "string"
          }
        },
        "required": [
          "addresses",
          "matchtype",
          "recommendationCode"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponseContainerIDS": {
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "matchDate": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponseIDS"
          },
          "status": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError"
            }
          }
        },
        "required": [
          "apiVersion",
          "matchDate",
          "response",
          "status",
          "errors"
        ]
      },
...
    }
  },
...
  "openapi": "3.0.0",
  "info": {
    "title": "ONS Address Index API",
    "description": "The ONS Address Index API lets you retrieve the official version of addresses in England and Wales. The data is currently updated on a 6 week rolling basis.",
    "version": "see /version below"
  }
}